Overhaul tabs UI
This completely redesigns the way the tabs UI is rendered based on
testing it on the phone with Sway at scale 2.

Individual sizes and paddings are now easier to modify to tweak things
to the correct sizes and the tabs' font size can be easily adjusted
separately from the URI bar text size.

Generally elements have been made bigger and specifically for the tab
close button the padding has been included in the touch hitbox to avoid
accidentally opening the tab when the intention was to close it.

Closes #6.