Change active tab closing behavior
Previously when closing the active tab, Kumo would always focus the
first tab next. This is a big jump and can lead to losing track of the
active tab when in the tab overview.

Instead Kumo now switches to the tab immediately before the closed tab,
which should keep the position of the active tab in the overview
consistent.

Closes #9.