Fix startup with scale sent before size
This fixes an issue where Kumo would crash in debug mode due to an
underflow when a compositor sent the fractional scale before the first
size configure.

To fix this all UI elements are now initialized to the default window
size on startup and size/scale updates are discarded if they did not
change.