Add previous page button
This adds a button for navigating through the engine's page history.

Navigation is only possible backwards, since that should be the primary
means of history navigation. That way the impact on available screen
real estate is minimal.

Forward navigation can be added later through a menu when long-pressing
on the "previous page" button.