Switch browser UI to primary surface
This fixes a regression introduced in 50bfcd5 where Kumo would not
render the first frame in Sway and thus stall indefinitely. This was
caused by the UI delivering the first frame, which wasn't counted by
Sway since it's a subsurface and the parent surface was not mapped yet.

Since the UI surface has changed to be a window-sized surface that is
always able to render in 50bfcd5 anyway, it should not be an issue to
just make it the primary surface instead.