Add IME support to browser engine
Previously the Wayland text-input protocol was only hooked up to the
text input in the UI. This patch extends this to the browser engine
using WebKit's `set_input_method_context` function.

This provides both the ability to submit text through IME and
automatically show/hide the input method. Preedit is still not supported
in the engine, since WebKit doesn't provide a way to submit preedit
text.

Closes #21.