Fix WPE activity state changes
This fixes an issue with the `set_focus` call introduced in b813f1f,
which updated the activity state to indicate focus but caused the engine
to assume the window wasn't focused.

This patch sets the activity state at startup to include all activity
states, then toggles the focus state whenever `set_focus` is called.