Fix engine overlapping UI surface
This fixes an issue with the viewporter logic of the engine surface
where the engine's surface wasn't correctly clamped to the maximum
desired engine size until a new engine frame is submitted.

The engine's viewporter source size is now set to ensure its surface is
never bigger than the UI surface.

See #20.