Fix slow rendering of new tabs
This fixes an issue where after opening a new tab, the engine UI would
not be updated until a new buffer was submitted by the engine. This is
problematic since the previous buffer will still be rendered without
being clamped to its desired dimensions, possibly exceeding the window
boundaries.

To make the browser feel more responsive and ensure that the engine
surface is never bigger than the UI surface, opening a new tab now
causes the engine's surface to be cleared. Since the UI background is
visually equivalent to the initial browser engine buffer, this looks to
the user like a new buffer was submitted instantly.

Closes #20.