Add autocomplete for URI bar
This patch adds in-memory tracking for visited pages, suggesting
previously visited websites as autocomplete hints which can be completed
using TAB.

While this data is shared between windows, it is currently not persisted
at all. This can likely be added in the future in combination with
setting config/data/cache directories for WebKit.