Fix URI bar scroll offset after redirect
This fixes an issue where the URI bar's scroll offset would not reset
when the URI changed completely. As a result, the scroll offset could be
outside of the maximum range allowed for the current input string.

To fix this, the scroll offset is now always reset when the text is
replaced.

Closes #39.