Fix scroll offset on backspace
This fixes an issue where the scroll offset would start showing empty
space behind the URI if the cursor wasn't at the end of the string when
using backspace.

The new implementation ensures to always clamp the scroll offset, which
should prevent empty space from showing up at the end unless the scroll
offset is already 0.