Fix URI content selection on initial click
This fixes an issue where the initial selection of the URI was
immediately cancelled on touch/pointer release due to the interpretation
as a tap action.

To fix this the focus status is now passed to the touch state
controller, entering a unique `TouchAction::Focus` state which ignores
touch input on the URI bar until it received focus.

Closes #41.