Fix inconsistent key repeat rate
This fixes an issue with key repetition where the interval between key
repeats was inconsistent since the interval between them was measured
from the end of the last repeat, rather than its beginning.

To ensure that the repeats are spaced as evenly as possible, GLib's
event source is now used to generate repeated events rather than
manually rescheduling the timer every repeat.

Closes #40.