Fix option menu creation performance
This fixes an issue during the creation of the option menu where a
significant amount of time was spent on repeated height calculation for
each individual option menu layout (~0.5ms per menu entry).

Since we only render a single line anyway and the line height is stored
as part of the font metrics, there's now a new `TextLayout::line_height`
function which calculates and caches the line height for each individual
font size.

This reduces the performance impact of `OptionMenu::scroll` with 50
elements from ~12ms to ~0.1ms.

See #42.