Optimize history menu rendering
This patch makes several optimizations to improve the performance of
opening and modifying the history menu popup.

Since d536a76 has introduced a cheap way to calculate item height
without rasterization, it was made possible to skip rendering of any
option menu item that is fully invisible.

The history menu specifically also has the property of often showing
very similar entries due to the minimal changes generally made during
text input. To optimize for this, a function to update the items in an
option menu has been introduced. This new mechanism will rely on a cache
of the last 100 items to avoid re-creating items that have recently been
seen in the menu.

Lastly, a function to show and hide option menus was introduced. The
history menu previously was frequently closed whenever there were no
suggestions, which would clear the cache introduced in this patch. To
keep the cache around, the history menu is now reused and its visibility
is used to show/hide it when necessary.

This also fixes a minor bug where `OptionMenu::scroll` would set
dirtiness to the result of `self.scroll_offset != scroll_offset`,
instead of assigning it only when `true`.

Closes #42.