Fix autosuggestions with empty path segments
This fixes an issue where odd autosuggestions would be suggested when
the input URI contained multiple successive slashes (for example
`example.org///`).

URIs are now normalized when storing them in the history, which avoids
bloating up the history with identical URIs that just differ due to
multiple successive slashes. As a result these URIs are never suggested
to the user, even if explicitly entered as such previously. However this
should never be necessary.

One caveat with this approach is that we might store an incorrect title
in the history. The following two URIs do not resolve to the same
resource for example:
 - https://example.org/test/  — "Example Domain"
 - https://example.org/test// — "404 - Not Found"

Closes #44.