Fix logical/physical conversion in option menu
This fixes an issue in the option menu where sizes were converted from
physical to logical and back, leading to some loss due to rounding in
the process.

Since the basis of most calculations for the option menu is the physical
line height, which can be uneven despite being at scale 2, it is best to
just do everything in physical and convert to logical at the last
possible moment.

Closes #43.