Fix history menu scale changes
This fixes an issue where the history menu would keep around incorrectly
scaled items with an incorrect position after the scale of the window
changed.

The original code used `close_history_menu` in the `configure` handling
to avoid worrying about this, however that got broken in bdcf5d4 since
it changed all related code to just hide the menu temporarily.

Since performance shouldn't be a big factor for a scale change and a
full re-rasterization of all text is necessary anyway, the original
approach has been reimplemented by replacing the `close_history_menu`
call with a full destruction of the cached history menu.