Fix direct scanout for toolbar surface
This fixes an issue where the main UI surface containing the URI bar
would never get scanned out on a PPP, since its overlay formats do not
support the z-index property and the surface overlapped with the engine
surface.

A new surface has been added for the sole purpose as functioning as the
engine backdrop, to ensure we can have both a backdrop and direct
scanout for the UI surface. This is uses the single-pixel buffer
protocol when available.

Closes #37.