Remove session when clearing additional windows
Previously sessions were only cleared when a tab was explicitly closed.
However when an entire window is manually closed it's generally
unexpected to see that window coming back the next time the browser is
opened.

To behave more predictably, Kumo now clears sessions when closing
windows as long as it is not the last remaining window. This allows
either restoring multiple windows after a `kill kumo`, or just restoring
your main browser window every time you open and close the app manually.

Closes #52.