Add whitelist-based cookie retention policy
This changes the way cookies are handled by adding a new SQLite
whitelist and deleting every cookie on startup which isn't part of the
whitelist.

Whitelisted domains can be managed by long-pressing on the document of a
page and selecting "Add/Remove Cookie Exception".

This patch nukes all existing sessions and browser history, since the
previous `history.sqlite` has been removed in favor of a new
`storage.sqlite` file which stores sessions, history, and cookie
exceptions.