Add tab groups
This patch adds tab groups, which allows bundling a collection of tabs
which all share the same network session. Each tab group shares local
storage and cookies and is isolated from all other tab groups. On
session recovery, the existing data shared between tabs in a group is
recovered.

Each tab group can also have an `ephemeral` flag set through the UI. If
the flag is set and the browser is closed, the tab group's shared data
will be deleted and none of the tabs are recovered. New tabs opened in a
group marked as ephemeral will also not be added to the history.