Add label to tab groups
This adds a little label at the top for each tab group, to help users
identify which tab group they're currently in.

This patch does not yet give a way to change this label, so it will
always be `Default` for the standard group and their UUID for all
others.

Since this now renders something above the tabs at the top, it also
fixes the issue where the tabs would render partially under the
persistent/new tab group buttons.

See #57.