Add button to close tab groups
This adds a dedicated button which allows closing tab groups regardless
of the number of tabs in them. Tab groups are no longer closed
automatically when the last tab in the group is closed.

To avoid cluttering the UI with more buttons, this button just replaces
the "new tab group" button while editing the group label, making it
relatively simple to delete tab groups while staying out of the way
during general use.

When a non-empty tab group is closed, all tabs inside it are
automatically discarded.

Closes #59.