Fix history item selection
This fixes a bug with the history option menu where clicking on an item
would close the popup rather than selecting that item.

This regression was likely introduced by c4317bf, which introduced
explicit keyboard focus for overlay popups. Since keyboard focus loss on
the URI bar automatically closes the history, it would close it before
the item itself was opened.

To ensure the URI input keeps being focused while interacting with the
history menu, we now no longer transfer focus to the Overlay unless the
popup which received the "touch down" event accepts keyboard input.

Closes #63.