Recover focused tab on session restore
This patch adds a new `focused` field to the `session` SQLite table
which tracks the tab that is currently focused.

No migration was written since SQLite doesn't provide a simple solution
for `ADD COLUMN IF NOT EXISTS`, so the `storage.sqlite` must be deleted
to update.

If the session table contains multiple focused tabs for a single window,
the last tab and its tab group will be focused.

Closes #64.