Add browser history UI
This patch adds a new UI which can be opened by pressing the three dots
in the top-left of the tabs view, allowing management of the browser
history.

The menu button is only available in the default tab group, since it
takes the place of the ephemeral mode button. Currently it directly
opens the browser history UI, but in the future it could open an option
menu to open different UIs for browser internals like settings.

The new history UI shows all entries chronologically, allowing users to
either open them in a new tab or removing entries from the history
entirely. Only a single entry will show up per URL, ordered using its
latest access timestamp.

Closes #67.