Add database migration infrastructure
This adds a new `db_version` table to the `storage.sqlite` file, which
allows detecting the state of the database file to enable more advanced
migrations without having to delete the entire file.

If the reported `db_version` is newer than supported by the executable,
Kumo will exit with an error to ensure no data is accidentally lost.

The database migrations themselves are run inside a transaction, to
ensure that no update is partially applied. On failure the database
stays unchanged and an error is emitted.