Add Servo engine support
This patch adds rudimentary support for the Servo browser engine as an
alternative to WPE WebKit. This makes it possible to compile Kumo
without needing to separately compile a patched WPE WebKit.

In its current state this is not functional enough to be useful, since
Servo does not provide an API that can reliably feed an event loop and
redraw the browser engine.

Many features that are theoretically supported by Servo are also not
implemented yet, since the event processing/rendering issues should be
sorted out first.