Fix fullscreen regressions
This patch fixes a regression which caused the main UI input URI to
render above the engine's surface while in fullscreen mode. While the UI
was not redrawn, the surface was not explicitly detached which caused it
to still be visible.

Additionally the WebKit fullscreen request handling system only
propagated fullscreen changes when requested by the browser, causing it
to get stuck when fullscreen mode was disabled from the compositor side.

Closes #83.