Request frame events for all surfaces
This works around an issue on Sway where frame events were frequently
omitted, since the toplevel surface is usually hidden by the engine
subsurface.

It was attempted to only send a frame request for one visible surface
that was updated this frame, but that also turned out not to work on
Sway.

To avoid excessive redraws, a combination of serial and time are used as
a heuristic.