Fix frame event commits
In 8ad0f39 a workaround was introduced to fix rendering on Sway, which
was incorrectly making the assumption that frame events would not be
sent unless a surface had a new buffer attached. Instead, the original
patch failed to properly add commits for these frame events before the
toplevel commit.

This patch removes the hacky workaround that mostly worked by
coincidence and instead replaces it with a mechanism that always picks
one surface that is visible and then sends a frame and commit for just
that surface.