Add features to app metainfo
This adds the list of features to the application metainfo, making it
easier for users to find out what sets Kumo apart.

The summary has also been changed is several places to drop the
redundant `A …` prefix and add a `Web` to make it explicit what kind of
browser Kumo is.