Add load progress highlight to tabs view
This patch adds a red accent indicating the progress being made while
loading a tab in the tab view, allowing the user to track progress when
loading multiple tabs at the same time.

To avoid showing a colored accent all the time, the progress bar
disappears as soon as it reaches 100% load progress.

Closes #84.