Fix color rounding
This fixes an issue with color inconsistencies between Cairo and OpenGL.
The default background color `0.1` translates to an 8 bit value of
`25.5`, which OpenGL corectly rounds to `26` while Cairo rounds it down
to `25`.

Since all colors are ultimately converted to 8 bit colored textures,
this patch changes the input colors to use `u8` instead of `f64` which
should prevent issues with rounding issues due to floating point values
which do not fit the render format of 8 bit colors.