Increase tab button input area
In 4b84450 the input area calculation for the URI buttons was changed to
use only the button sizes. While technically this is correct, it
struggles a bit with input precision on mediocre touchscreens.

This patch restores that functionality to include parts of the padding
in the button's input geometry, making it easier for users to hit these
buttons.