Add page text search
This patch hooks up a new UI to WebKit's `FindController`, allowing
users to search for text on websites.

The default WebKit colors aren't ideal, which is why this was paired
with the 8b99394 WebKit patch.

The UI reuses the existing URI input bar, but changes its size a bit and
adds buttons for navigating between matches and closing the search.

Closes #100.