Fix favicon backgrounds
This fixes an issue introduced in aeb1e01 where favicons not fully
filling the favicon rectangle would show a black rectangular outline,
since the tab had a cutout for favicons which was not cleared with the
tab's background color.

To still allow the audio playback indicator to be rendered above the
favicon while rendering the favicon on top of a properly cleared tab
background, the audio icon is now rendered and cached separately after
both the tab and the favicon are rendered.