Ignore failed websites for browser history
This uses a heuristics-based approach to determine which page load
events contributed to a successful page load and only updates the page
history when a successful load is performed.

With this approach, broken pages should never be added to the history.
While it might be possible that some successful loads aren't added
correctly, that should be very rare and not impact user experience.

Closes #62.