Fix unreadable text for error pages
This fixes an issue where error pages would always use black text,
regardless of the engine's background color, leading to invisible text
with black backgrounds.

While overriding the CSS of builtin pages is not possible in WebKit,
it's possible to ship custom error pages. This patch adds a simple error
page with templating to display errors with custom layout and colors.