Add button to reset config to its defaults
Previously the config UI had two buttons: A checkmark and an X. It was
not entirely clear what each individual button does, but one persisted
the config to the file while the other just quit the UI. This behavior
is kept, but the icons have been switched to a floppy and the left
arrow, to make their behavior mor clear.

Additionally this adds a new third button which allows reverting the
config to the default. When changes are made that are not desired by the
user, it's unlikely that they want to just leave the UI or manually
revert these changes. To ensure failed experiments are easy to recover
from, this reset button reverts all changes back to the ones currently
stored in the configuration file.

To indicate when changes are present, save and reset icons are now
hidden while all settings are at their default. This also provides
better feedback to the user that the save was successful.